% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fly.R
\name{fly}
\alias{fly}
\title{Start a run on HTCondor}
\usage{
fly(conditions, rscript, reps, input_files = NULL, sleep_time = 1,
  log = "log.$(Process)", output = "out.$(Process)",
  error = "error.$(Process)", universe = "vanilla", requirements = NULL,
  other = NULL, submit_file_name = "Rjob.submit")
}
\arguments{
\item{conditions}{Grid of conditions to use for run}

\item{rscript}{The R file to execute}

\item{reps}{Number of replicates to do}

\item{input_files}{Any other files needed to run}

\item{sleep_time}{How long between calls to condor_submit}

\item{log}{Log file}

\item{output}{Output file (in addition to ones the script itself creates)}

\item{error}{Error file}

\item{universe}{HTCondor universe to use}

\item{requirements}{Requirements to pass as a requirements call}

\item{other}{Other lines to add to the submit file (such as rank)}

\item{submit_file_name}{Name of the final submit file}
}
\description{
This takes a data.frame of arguments to use for your actual run, often generated by something like \code{expand.grid} so that each row represents a unique combination of conditions. If you want each condition to be run more than once, change \code{reps}. It expects the \code{rscript} to be an R file that takes unnamed input arguments, with the first argument being the job number.
}
\seealso{
packages in the High Performance Task View for R.
}
