% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_file_create.R
\name{submit_file_create}
\alias{submit_file_create}
\title{Create the submit file for HTCondor}
\usage{
submit_file_create(input, log = "log.$(Process)", output = "out.$(Process)",
  error = "error.$(Process)", universe = "vanilla", requirements = NULL,
  other = NULL, executable_file_name = "run.sh",
  submit_file_name = "Rjob.submit")
}
\arguments{
\item{input}{Input files}

\item{log}{Log file}

\item{output}{Output file (in addition to ones the script itself creates)}

\item{error}{Error file}

\item{universe}{HTCondor universe to use}

\item{requirements}{Requirements to pass as a requirements call}

\item{other}{Other lines to add to the submit file (such as rank)}

\item{executable_file_name}{Executable file name to make}

\item{submit_file_name}{Name of the final submit file}
}
\description{
This creates the submit file run with \code{condor_submit}. The default arguments take advantage of HTCondor macros to append process id to each output. This has named arguments for usual submit file elements; if you want others, add them to the \code{other} argument in the form you want them in a typical submit file.
}
\examples{
\dontrun{
 submit_file_create(input=c("birds.tre", "bird_traits.csv"), executable_file_name="batch.R"))
}
}
