% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_file_create.R
\name{submit_file_create}
\alias{submit_file_create}
\title{Create the submit file for HTCondor}
\usage{
submit_file_create(input, reps = 1, executable = "run.sh",
  log = "log.$(Process)", output = "out.$(Process)",
  error = "error.$(Process)", universe = "vanilla", requirements = NULL,
  other = NULL, submit_file_name = "Rjob.submit")
}
\arguments{
\item{input}{Input files}

\item{reps}{Number of replicates to do}

\item{executable}{Executable file}

\item{log}{Log file}

\item{output}{Output file (in addition to ones the script itself creates)}

\item{error}{Error file}

\item{universe}{HTCondor universe to use}

\item{requirements}{Requirements to pass as a requirements call}

\item{other}{Other lines to add to the submit file (such as rank)}

\item{submit_file_name}{Name of the final submit file}
}
\description{
Create the submit file for HTCondor
}
